Straight deck rendering
